pyproject.toml
[project]
name = "leetcodes"
version = "0.1.0"
description = "Leetcodes."
readme = "README.md"
authors = [
    { name = "the-citto", email = "alex.citterio@hotmail.co.uk" }
]
requires-python = ">=3.13"
dependencies = []

[project.scripts]
# leetcodes = "leetcodes:cli.cli"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipython>=9.8.0",
]
test = [
    "flake8-pyproject>=1.2.4",
    "google-benchmark>=1.9.4",
    "pytest-checkers[all]>=1.0.3",
    "pytest-cov>=7.0.0",
    "tox>=4.32.0",
    "tox-uv>=1.29.0",
]


[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
exclude = [
    ".git", 
    ".venv",
    "__pycache__",
    ".tox",
    "*.pyi",
]
ignore = ["E203"]

[tool.isort]
profile = "black"
multi_line_output = 3
force_grid_wrap = 2

[tool.mypy]
strict = true
mypy_path = ["src", "typesheds"]
disallow_untyped_defs = true
exclude = ["build/*"]

[tool.pyright]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
stubPath = "typesheds"
ignore = [".venv/*"]

[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = [ "ALL" ]
ignore = [
    "D203", # incorrect-blank-line-before-class
    # "D211", # blank-line-before-class
    "D213", # multi-line-summary-second-line
    # "D212", # multi-line-summary-first-line
    "ERA001", # commented-out code
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN401",   # any-type
    "PLR2004",  # magic-value-comparison
    "S101",     # assert
]
"benchmarks/**/*.py" = [
    "S311",
]

# [tool.ty.environment]
# root = ["src", "tests", "benchmarks"]
[tool.ty.rules]
invalid-type-arguments = "ignore"
unsupported-operator = "ignore"
possibly-missing-attribute = "ignore"

